requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
PyMySQL==1.1.0
python-dotenv==1.0.0
bcrypt==4.1.2
Werkzeug==3.0.1
WTForms==3.1.1
email-validator==2.1.0
opencv-python>=4.12.0
ultralytics>=8.0.0
scipy>=1.9.0
fastdtw>=0.3.4
Pillow>=9.0.0
numpy>=2.0.0